Added helper functions to nixnet_utilties
